fix docs

+++ var/spack/repos/builtin/packages/charmpp/package.py
@@ -18,10 +18,10 @@ class Charmpp(Package):
     (your laptop) to the largest supercomputers."""
 
     homepage = "http://charmplusplus.org"
-    url      = "https://charm.cs.illinois.edu/distrib/charm-6.8.2.tar.gz"
+    url      = "https://github.com/UIUC-PPL/charm/archive/v6.10.0.tar.gz"
     git      = "https://github.com/UIUC-PPL/charm.git"
 
-    version("develop", branch="charm")
+    version("develop", branch="master")
     version('6.10.0', sha256='7c526a78aa0c202b7f0418b345138e7dc40496f0bb7b9e301e0381980450b25c')
     version("6.9.0", sha256="85ed660e46eeb7a6fc6b32deab08226f647c244241948f6b592ebcd2b6050cbd")
     version("6.8.2", sha256="08e6001b0e9cd00ebde179f76098767149bf7e454f29028fb9a8bfb55778698e")
@@ -263,13 +263,14 @@ class Charmpp(Package):
                         pass
         shutil.rmtree(join_path(prefix, "tmp"))
 
-        # A broken 'doc' link in the prefix can break the build.
-        # Remove it and replace it if it is broken.
-        try:
-            os.stat(prefix.doc)
-        except OSError:
-            os.remove(prefix.doc)
-            mkdirp(prefix.doc)
+        if self.spec.satisfies('@6.9.99'):
+            # A broken 'doc' link in the prefix can break the build.
+            # Remove it and replace it if it is broken.
+            try:
+                os.stat(prefix.doc)
+            except OSError:
+                os.remove(prefix.doc)
+                mkdirp(prefix.doc)
 
     @run_after('install')
     @on_package_attributes(run_tests=True)
